deleted unimportant files and sent the response in one line Relates #30 #38 #41
