+++ public/js/dom.js
@@ -4,7 +4,12 @@ const lyrics = document.querySelector('.content__text');
 
 search.addEventListener('click', (e) => {
   e.preventDefault();
-  fetchApi((obj) => {
-    lyrics.textContent = obj.message.body.lyrics.lyrics_body;
-  });
+  fetch('/song', {
+    method: 'POST',
+    body: JSON.stringify({ name: song.value }),
+    headers: { 'Content-Type': 'application/json' },
+  })
+    .then(res => res.json())
+    .then(data => (lyrics.textContent = data.message.body.lyrics.lyrics_body))
+    .catch(error => console.log(error));
 });
