--- public/js/logic.js
@@ -1,6 +1,0 @@
-const fetchApi = (cb) => {
-  fetch('/api')
-    .then(res => res.json())
-    .then(data => cb(data))
-    .catch(error => console.log(error));
-};
